refactor(tests): extract helper functions from the tests
Extract helper functions from the tests
