+++ src/Components/mocks/index.js
@@ -1,0 +1,21 @@
+const data = [
+  {
+    title: 'Avengers',
+    events: [
+      {
+        title: 'Avengers:StartGame',
+        subtitle: 'StartGame',
+        content: 'content',
+        location: 'Knowhere',
+        label: '#avengers #marvel',
+        img: {
+          url:
+            'https://images.unsplash.com/photo-1600790078201-5490baf711d6?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=500&q=60',
+          alt: 'picture',
+        },
+      },
+    ],
+  },
+];
+
+export default data;
